BalanceError module: change report way to email instead of opening Contact module
- remove Contact module
