--- UnstoppableWallet/UnstoppableWallet/Modules/Contact/ContactInteractor.swift
@@ -1,26 +1,0 @@
-class ContactInteractor {
-    private let appConfigProvider: IAppConfigProvider
-    private let pasteboardManager: IPasteboardManager
-
-    init(appConfigProvider: IAppConfigProvider, pasteboardManager: IPasteboardManager) {
-        self.appConfigProvider = appConfigProvider
-        self.pasteboardManager = pasteboardManager
-    }
-
-}
-
-extension ContactInteractor: IContactInteractor {
-
-    var email: String {
-        appConfigProvider.reportEmail
-    }
-
-    var telegramWalletHelpAccount: String {
-        appConfigProvider.telegramAccount
-    }
-
-    func copyToClipboard(string: String) {
-        pasteboardManager.set(value: string)
-    }
-
-}
