--- UnstoppableWallet/UnstoppableWallet/Modules/Contact/ContactRouter.swift
@@ -1,62 +1,0 @@
-import UIKit
-import MessageUI
-
-class ContactRouter: NSObject {
-    weak var viewController: UIViewController?
-}
-
-extension ContactRouter: IContactRouter {
-
-    var canSendMail: Bool {
-        MFMailComposeViewController.canSendMail()
-    }
-
-    func openSendMail(recipient: String) {
-        let controller = MFMailComposeViewController()
-        controller.setToRecipients([recipient])
-        controller.mailComposeDelegate = self
-
-        viewController?.present(controller, animated: true)
-    }
-
-    func openTelegram(account: String) {
-        guard let appUrl = URL(string: "tg://resolve?domain=\(account)") else {
-            return
-        }
-
-        if UIApplication.shared.canOpenURL(appUrl) {
-            UIApplication.shared.open(appUrl)
-        } else if let webUrl = URL(string: "https://t.me/\(account)") {
-            UIApplication.shared.open(webUrl)
-        }
-    }
-
-    func showDebugLog() {
-        viewController?.navigationController?.pushViewController(DebugRouter.module(), animated: true)
-    }
-
-}
-
-extension ContactRouter: MFMailComposeViewControllerDelegate {
-
-    func mailComposeController(_ controller: MFMailComposeViewController, didFinishWith result: MFMailComposeResult, error: Error?) {
-        controller.dismiss(animated: true)
-    }
-
-}
-
-extension ContactRouter {
-
-    static func module() -> UIViewController {
-        let router = ContactRouter()
-        let interactor = ContactInteractor(appConfigProvider: App.shared.appConfigProvider, pasteboardManager: App.shared.pasteboardManager)
-        let presenter = ContactPresenter(interactor: interactor, router: router)
-        let viewController = ContactViewController(delegate: presenter)
-
-        presenter.view = viewController
-        router.viewController = viewController
-
-        return viewController
-    }
-
-}
